devices: docker_factory: fixes for using remote docker host to create containers

+++ boardfarm/devices/docker_factory.py
@@ -2,6 +2,7 @@ import pexpect
 import sys
 import os
 import atexit
+import re
 
 import linux
 
@@ -16,6 +17,7 @@ class DockerFactory(linux.LinuxDevice):
     created_docker = False
     extra_devices = []
     target_cname = []
+    created_network = True
 
     def __str__(self):
         return self.name
@@ -28,14 +30,16 @@ class DockerFactory(linux.LinuxDevice):
         self.ipaddr = kwargs.pop('ipaddr', None)
         self.iface = kwargs.pop('iface', None)
         self.docker_network = kwargs.pop('docker_network', None)
-        self.env = kwargs.pop('env', None)
+        env = self.env = kwargs.pop('env', None)
         self.name = kwargs.pop('name')
         self.cname = self.name + '-${uniq_id}'
+        self.username = kwargs.pop('username', 'root')
+        self.password = kwargs.pop('password', 'bigfoot1')
 
         if self.ipaddr is not None:
             # TOOO: we rely on correct username and key and standard port
             pexpect.spawn.__init__(self, command="ssh",
-                                       args=['%s' % (self.ipaddr),
+                                       args=['%s@%s' % (self.username, self.ipaddr),
                                              '-o', 'StrictHostKeyChecking=no',
                                              '-o', 'UserKnownHostsFile=/dev/null',
                                              '-o', 'ServerAliveInterval=60',
@@ -47,12 +51,21 @@ class DockerFactory(linux.LinuxDevice):
         if 'BFT_DEBUG' in os.environ:
             self.logfile_read = sys.stdout
 
-        self.expect(pexpect.TIMEOUT, timeout=1)
+        # TODO: reused function for ssh auth for all of this
+        if 0 == self.expect(['assword', pexpect.TIMEOUT], timeout=10):
+            self.sendline(self.password)
+
         self.sendline('export PS1="docker_session>"')
         self.expect(self.prompt)
         self.sendline('echo FOO')
         self.expect_exact('echo FOO')
         self.expect(self.prompt)
+
+        if self.ipaddr != 'localhost':
+            print env
+            for k, v in env.iteritems():
+                self.sendline('export %s=%s' % (k, v))
+                self.expect(self.prompt)
 
         self.set_cli_size(200)
 
@@ -62,13 +75,16 @@ class DockerFactory(linux.LinuxDevice):
             self.expect(self.prompt)
             if 'error fetching interface information: Device not found' not in self.before:
                 break
+            self.expect(pexpect.TIMEOUT, timeout=5)
 
         # iface set, we need to create network
         if self.iface is not None:
             self.sendline('docker network create -d macvlan -o parent=%s -o macvlan_mode=bridge %s' % (self.iface, self.cname))
             self.expect(self.prompt)
             assert 'Error response from daemon: could not find an available, non-overlapping IPv4 address pool among the defaults to assign to the network' not in self.before
-            assert ' is already using parent interface ' not in self.before
+            if ' is already using parent interface ' in self.before:
+                self.cname = re.findall('dm-(.*) is already', self.before)[0]
+                self.created_network = False
             self.sendline('docker network ls')
             self.expect_exact('docker network ls')
             self.expect(self.prompt)
@@ -120,7 +136,7 @@ class DockerFactory(linux.LinuxDevice):
         self.clean_docker_network()
 
     def clean_docker_network(self):
-        if self.created_docker_network == True:
+        if self.created_docker_network == True and self.created_network == True:
             self.sendline('docker network rm %s' % self.cname)
             self.expect(self.prompt)
             self.sendline('docker network ls')
